pyproject.toml
[tool.poetry]
name = "autofl"
version = "0.2.0"
description = ""
authors = ["Cezar Sas <cezar.sas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.104.1"
gunicorn = "^21.2.0"
uvicorn = "^0.24.0"
pandas = "^2.1.2"
hydra-core = "^1.3.2"
setuptools = "^68.2.0"
multiset = "^3.0.1"
scikit-learn = "^1.3.2"
pydantic = "^2.4.2"
gitpython = "^3.1.40"
loguru = "^0.7.1"
tqdm = "^4.66.1"
yake = { git = "git@github.com:LIAAD/yake.git" }
python-rake = "^1.5.0"
more-itertools = "^10.1.0"
tree-sitter = "^0.20.2"
sqlalchemy = "^2.0.21"
psycopg = {extras = ["binary"], version = "^3.1.12"}
gensim = "^4.3.2"
fasttext-wheel = "^0.9.2"
transformers = "^4.35.0"
sentence-transformers = "^2.2.2"

[tool.poetry.group.dev.dependencies]
notebook = "^6.5.4"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
